requirements.txt
asgiref==3.7.2
boto3==1.34.84
botocore==1.34.84
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.2
django-allauth==0.61.1
django-bootstrap5==24.1
django-environ==0.11.2
django-storages==1.14.2
fonttools==4.51.0
gunicorn==21.2.0
html5lib==1.1
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.9.0
PyJWT==2.8.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.3.0
typing_extensions==4.9.0
urllib3==2.2.1
weasyprint==61.2
webencodings==0.5.1
whitenoise==6.6.0
zopfli==0.2.3
